requirements.txt
flask
pandas
faiss-cpu
numpy
sentence-transformers
streamlit